extra layer of refactor into LoadedPlaylist
