+++ src/App/Playlist/LoadedPlaylist/Main/Main.scss
@@ -1,2 +1,3 @@
 .playlist-main-container {
+  width: 100vw;
 }
