+++ src/App/Playlist/LoadedPlaylist/Main/PlaylistHeader/PlaylistHeader.scss
