+++ src/App/Playlist/LoadedPlaylist/Main/Tracklist/TrackContainer/AlbumCover/AlbumCover.scss
