+++ src/App/Playlist/LoadedPlaylist/Main/Tracklist/TrackContainer/MainInfo/MainInfo.scss
