+++ src/App/Playlist/LoadedPlaylist/Main/Tracklist/TrackContainer/ProgressBar/ProgressBar.scss
