+++ src/App/Playlist/LoadedPlaylist/Sidebar/Sidebar.scss
@@ -1,0 +1,10 @@
+.sidebar {
+  // should appear over main view
+  z-index: 1;
+  position: absolute;
+  height: 100%;
+  width: 120px;
+  background: white;
+  box-shadow: 0 -4px 13px 2px rgba(50, 50, 93, 0.25),
+    0 8px 16px -8px rgba(0, 0, 0, 0.3);
+}
