--- src/App/Playlist/Sidebar/Sidebar.scss
@@ -1,2 +1,0 @@
-.sidebar {
-}
